+++ tests/unit/Spoke/Liquidations/Spoke.LiquidationCall.BadDebt.t.sol
@@ -291,7 +291,7 @@ contract LiquidationCallCloseFactorBadDebtTest is SpokeLiquidationBase {
     uint256 skipTime,
     uint256 desiredHf
   ) public {
-    liqConfig.closeFactor = uint128(HEALTH_FACTOR_LIQUIDATION_THRESHOLD);
+    liqConfig.closeFactor = HEALTH_FACTOR_LIQUIDATION_THRESHOLD;
     test_liquidationCall_fuzz_badDebt(
       collateralReserveId,
       debtReserveId,
